+++ rog-control-center/src/widgets/anime_power.rs
@@ -1,4 +1,5 @@
 use egui::{RichText, Ui};
+use rog_anime::usb::Brightness;
 use rog_platform::supported::SupportedFunctions;
 
 use crate::system_state::SystemState;
@@ -7,42 +8,50 @@ pub fn anime_power_group(_supported: &SupportedFunctions, states: &mut SystemSta
     ui.heading("AniMe Matrix Settings");
     ui.label("Options are incomplete. Awake + Boot should work");
 
-    let mut changed = false;
+    let mut brightness = states.anime.display_brightness as u8;
 
     ui.horizontal_wrapped(|ui| {
         ui.vertical(|ui| {
             let h = 16.0;
             ui.set_row_height(22.0);
             ui.horizontal_wrapped(|ui| {
-                ui.label(RichText::new("Brightness").size(h));
+                ui.label(RichText::new("Display brightness").size(h));
             });
             ui.horizontal_wrapped(|ui| {
-                ui.label(RichText::new("Boot").size(h));
+                ui.label(RichText::new("Display enabled").size(h));
             });
             ui.horizontal_wrapped(|ui| {
-                ui.label(RichText::new("Awake").size(h));
+                ui.label(RichText::new("Animations enabled").size(h));
             });
-            ui.horizontal_wrapped(|ui| {
-                ui.label(RichText::new("Sleep").size(h));
-            });
+            // ui.horizontal_wrapped(|ui| {
+            //     ui.label(RichText::new("Sleep").size(h));
+            // });
         });
         ui.vertical(|ui| {
             ui.set_row_height(22.0);
             ui.horizontal_wrapped(|ui| {
+                if ui.add(egui::Slider::new(&mut brightness, 0..=3)).changed() {
+                    states
+                        .asus_dbus
+                        .proxies()
+                        .anime()
+                        .set_brightness(Brightness::from(brightness))
+                        .map_err(|err| {
+                            states.error = Some(err.to_string());
+                        })
+                        .ok();
+                }
+            });
+            ui.horizontal_wrapped(|ui| {
                 if ui
-                    .add(egui::Slider::new(&mut states.anime.bright, 0..=254))
+                    .checkbox(&mut states.anime.display_enabled, "Enable")
                     .changed()
                 {
-                    changed = true;
-                }
-            });
-            ui.horizontal_wrapped(|ui| {
-                if ui.checkbox(&mut states.anime.boot, "Enable").changed() {
                     states
                         .asus_dbus
                         .proxies()
                         .anime()
-                        .set_animation_enabled(states.anime.boot)
+                        .set_enable_display(states.anime.display_enabled)
                         .map_err(|err| {
                             states.error = Some(err.to_string());
                         })
@@ -50,23 +59,26 @@ pub fn anime_power_group(_supported: &SupportedFunctions, states: &mut SystemSta
                 }
             });
             ui.horizontal_wrapped(|ui| {
-                if ui.checkbox(&mut states.anime.awake, "Enable").changed() {
+                if ui
+                    .checkbox(&mut states.anime.builtin_anims_enabled, "Enable")
+                    .changed()
+                {
                     states
                         .asus_dbus
                         .proxies()
                         .anime()
-                        .set_awake_enabled(states.anime.awake)
+                        .set_builtins_enabled(states.anime.builtin_anims_enabled)
                         .map_err(|err| {
                             states.error = Some(err.to_string());
                         })
                         .ok();
                 }
             });
-            ui.horizontal_wrapped(|ui| {
-                if ui.checkbox(&mut states.anime.sleep, "Enable").changed() {
-                    changed = true;
-                }
-            });
+            // ui.horizontal_wrapped(|ui| {
+            //     if ui.checkbox(&mut states.anime.sleep, "Enable").changed() {
+            //         changed = true;
+            //     }
+            // });
         });
     });
 }